nintendo/vsnes.cpp: Various cleanups.
- Merged vsnes_m.cpp and vsnes.h into vsnes.cpp.
- Split into base class and three separate classes for UniSystem, DualSystem, and SMB bootleg machines.
- Merged DualSystem's main/sub VROM banking functions.
